Trait hotshot_consensus::ConsensusApi

pub trait ConsensusApi<TYPES: NodeType, LEAF: LeafType<NodeType = TYPES>, I: NodeImplementation<TYPES>>: Send + Sync {

    fn total_nodes(&self) -> NonZeroUsize;
    fn propose_min_round_time(&self) -> Duration;
    fn propose_max_round_time(&self) -> Duration;
    fn store_leaf<'life0, 'async_trait>(
        &'life0 self,
        old_anchor_view: TYPES::Time,
        leaf: LEAF
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn max_transactions(&self) -> NonZeroUsize;
    fn min_transactions(&self) -> usize;
    fn should_start_round<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_direct_message<'life0, 'async_trait, PROPOSAL, VOTE>(
        &'life0 self,
        recipient: TYPES::SignatureKey,
        message: ConsensusMessage<TYPES, I>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
        VOTE: 'async_trait + VoteType<TYPES>,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_direct_da_message<'life0, 'async_trait, PROPOSAL, VOTE>(
        &'life0 self,
        recipient: TYPES::SignatureKey,
        message: ConsensusMessage<TYPES, I>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
        VOTE: 'async_trait + VoteType<TYPES>,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_broadcast_message<'life0, 'async_trait, PROPOSAL, VOTE>(
        &'life0 self,
        message: ConsensusMessage<TYPES, I>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
        VOTE: 'async_trait + VoteType<TYPES>,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_event<'life0, 'async_trait>(
        &'life0 self,
        event: Event<TYPES, LEAF>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn public_key(&self) -> &TYPES::SignatureKey;
    fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey;
    fn send_da_broadcast<'life0, 'async_trait>(
        &'life0 self,
        message: ConsensusMessage<TYPES, I>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn send_view_error<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time,
        error: Arc<HotShotError<TYPES>>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_replica_timeout<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_next_leader_timeout<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_decide<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time,
        leaf_views: Vec<LEAF>,
        decide_qc: QuorumCertificate<TYPES, LEAF>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_view_finished<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }

}

The API that HotStuff needs to talk to the system. This should be implemented in the hotshot crate and passed to all functions on HotStuff.

Required Methods

fn total_nodes(&self) -> NonZeroUsize

Total number of nodes in the network. Also known as n.

fn propose_min_round_time(&self) -> Duration

The minimum amount of time a leader has to wait before sending a propose

fn propose_max_round_time(&self) -> Duration

The maximum amount of time a leader can wait before sending a propose.ConsensusApi If this time is reached, the leader has to send a propose without transactions.

fn store_leaf<'life0, 'async_trait>(
&'life0 self,
old_anchor_view: TYPES::Time,
leaf: LEAF
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Store a leaf in the storage

fn max_transactions(&self) -> NonZeroUsize

Retuns the maximum transactions allowed in a block

fn min_transactions(&self) -> usize

Returns the minimum transactions that must be in a block

fn should_start_round<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns true if hotstuff should start the given round. A round can also be started manually by sending NewView to the leader.

In production code this should probably always return true.

fn send_direct_message<'life0, 'async_trait, PROPOSAL, VOTE>(
&'life0 self,
recipient: TYPES::SignatureKey,
message: ConsensusMessage<TYPES, I>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
VOTE: 'async_trait + VoteType<TYPES>,
Self: 'async_trait,
'life0: 'async_trait,

Send a direct message to the given recipient

fn send_direct_da_message<'life0, 'async_trait, PROPOSAL, VOTE>(
&'life0 self,
recipient: TYPES::SignatureKey,
message: ConsensusMessage<TYPES, I>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
VOTE: 'async_trait + VoteType<TYPES>,
Self: 'async_trait,
'life0: 'async_trait,

send a direct message using the DA communication channel

fn send_broadcast_message<'life0, 'async_trait, PROPOSAL, VOTE>(
&'life0 self,
message: ConsensusMessage<TYPES, I>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
PROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
VOTE: 'async_trait + VoteType<TYPES>,
Self: 'async_trait,
'life0: 'async_trait,

Send a broadcast message to the entire network.

fn send_event<'life0, 'async_trait>(
&'life0 self,
event: Event<TYPES, LEAF>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Notify the system of an event within hotshot-consensus.

fn public_key(&self) -> &TYPES::SignatureKey

Get a reference to the public key.

fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey

Get a reference to the private key.

fn send_da_broadcast<'life0, 'async_trait>(
&'life0 self,
message: ConsensusMessage<TYPES, I>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Send a broadcast to the DA comitee, stub for now

Provided Methods

fn send_view_error<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time,
error: Arc<HotShotError<TYPES>>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of an error

fn send_replica_timeout<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of a replica timeout

fn send_next_leader_timeout<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of a next leader timeout

fn send_decide<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time,
leaf_views: Vec<LEAF>,
decide_qc: QuorumCertificate<TYPES, LEAF>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

sends a decide event down the channel

fn send_view_finished<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Sends a ViewFinished event