Trait hotshot_consensus::traits::ConsensusApi

pub trait ConsensusApi<TYPES: NodeType, LEAF: LeafType<NodeType = TYPES>, PROPOSAL: ProposalType<NodeType = TYPES>>: Send + Sync {

    fn total_nodes(&self) -> NonZeroUsize;
    fn threshold(&self) -> NonZeroU64;
    fn propose_min_round_time(&self) -> Duration;
    fn propose_max_round_time(&self) -> Duration;
    fn store_leaf<'life0, 'async_trait>(
        &'life0 self,
        old_anchor_view: TYPES::Time,
        leaf: LEAF
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn max_transactions(&self) -> NonZeroUsize;
    fn min_transactions(&self) -> usize;
    fn make_vote_token(
        &self,
        view_number: TYPES::Time
    ) -> Result<Option<TYPES::VoteTokenType>, ElectionError>;
    fn get_leader<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = TYPES::SignatureKey> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn should_start_round<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_direct_message<'life0, 'async_trait>(
        &'life0 self,
        recipient: TYPES::SignatureKey,
        message: ConsensusMessage<TYPES, LEAF, PROPOSAL>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_broadcast_message<'life0, 'async_trait>(
        &'life0 self,
        message: ConsensusMessage<TYPES, LEAF, PROPOSAL>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn send_event<'life0, 'async_trait>(
        &'life0 self,
        event: Event<TYPES, LEAF>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn public_key(&self) -> &TYPES::SignatureKey;
    fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey;
    fn send_da_broadcast<'life0, 'async_trait, DAPROPOSAL>(
        &'life0 self,
        message: ConsensusMessage<TYPES, LEAF, DAPROPOSAL>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        DAPROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn is_valid_dac(
        &self,
        dac: &LEAF::DACertificate,
        block_commitment: Commitment<TYPES::BlockType>
    ) -> bool;
    fn is_valid_qc(&self, qc: &LEAF::QuorumCertificate) -> bool;
    fn is_valid_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        data: VoteData<TYPES, LEAF>,
        view_number: TYPES::Time,
        vote_token: Checked<TYPES::VoteTokenType>
    ) -> bool;
    fn accumulate_qc_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        leaf_commitment: Commitment<LEAF>,
        vote_token: TYPES::VoteTokenType,
        view_number: TYPES::Time,
        accumlator: CertificateAccumulator<TYPES::VoteTokenType, LEAF>
    ) -> Either<CertificateAccumulator<TYPES::VoteTokenType, LEAF>, QuorumCertificate<TYPES, LEAF>>;
    fn accumulate_da_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        block_commitment: Commitment<TYPES::BlockType>,
        vote_token: TYPES::VoteTokenType,
        view_number: TYPES::Time,
        accumlator: CertificateAccumulator<TYPES::VoteTokenType, TYPES::BlockType>
    ) -> Either<CertificateAccumulator<TYPES::VoteTokenType, TYPES::BlockType>, DACertificate<TYPES>>;

    fn is_leader<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_view_error<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time,
        error: Arc<HotShotError<TYPES>>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_replica_timeout<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_next_leader_timeout<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_decide<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time,
        leaf_views: Vec<LEAF>,
        decide_qc: LEAF::QuorumCertificate
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn send_view_finished<'life0, 'async_trait>(
        &'life0 self,
        view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn sign_da_proposal(
        &self,
        block_commitment: &Commitment<TYPES::BlockType>
    ) -> EncodedSignature { ... }
    fn sign_validating_or_commitment_proposal(
        &self,
        leaf_commitment: &Commitment<LEAF>
    ) -> EncodedSignature { ... }
    fn sign_da_vote(
        &self,
        block_commitment: Commitment<TYPES::BlockType>
    ) -> (EncodedPublicKey, EncodedSignature) { ... }
    fn sign_yes_vote(
        &self,
        leaf_commitment: Commitment<LEAF>
    ) -> (EncodedPublicKey, EncodedSignature) { ... }
    fn sign_no_vote(
        &self,
        leaf_commitment: Commitment<LEAF>
    ) -> (EncodedPublicKey, EncodedSignature) { ... }
    fn sign_timeout_vote(
        &self,
        view_number: TYPES::Time
    ) -> (EncodedPublicKey, EncodedSignature) { ... }
    fn create_da_message(
        &self,
        justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
        block_commitment: Commitment<TYPES::BlockType>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, LEAF, PROPOSAL> { ... }
    fn create_yes_message(
        &self,
        justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
        leaf_commitment: Commitment<LEAF>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, LEAF, PROPOSAL> { ... }
    fn create_no_message(
        &self,
        justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
        leaf_commitment: Commitment<LEAF>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, LEAF, PROPOSAL> { ... }
    fn create_timeout_message(
        &self,
        justify_qc: LEAF::QuorumCertificate,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, LEAF, PROPOSAL> { ... }

}

The API that HotStuff needs to talk to the system. This should be implemented in the hotshot crate and passed to all functions on HotStuff.

Required Methods

fn total_nodes(&self) -> NonZeroUsize

Total number of nodes in the network. Also known as n.

fn threshold(&self) -> NonZeroU64

The amount of stake required to reach a decision. See implementation of Election for more details.

fn propose_min_round_time(&self) -> Duration

The minimum amount of time a leader has to wait before sending a propose

fn propose_max_round_time(&self) -> Duration

The maximum amount of time a leader can wait before sending a propose.ConsensusApi If this time is reached, the leader has to send a propose without transactions.

fn store_leaf<'life0, 'async_trait>(
&'life0 self,
old_anchor_view: TYPES::Time,
leaf: LEAF
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Store a leaf in the storage

fn max_transactions(&self) -> NonZeroUsize

Retuns the maximum transactions allowed in a block

fn min_transactions(&self) -> usize

Returns the minimum transactions that must be in a block

fn make_vote_token(
&self,
view_number: TYPES::Time
) -> Result<Option<TYPES::VoteTokenType>, ElectionError>

Generates and encodes a vote token

fn get_leader<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = TYPES::SignatureKey> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns the I::SignatureKey of the leader for the given round and stage

fn should_start_round<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns true if hotstuff should start the given round. A round can also be started manually by sending NewView to the leader.

In production code this should probably always return true.

fn send_direct_message<'life0, 'async_trait>(
&'life0 self,
recipient: TYPES::SignatureKey,
message: ConsensusMessage<TYPES, LEAF, PROPOSAL>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Send a direct message to the given recipient

fn send_broadcast_message<'life0, 'async_trait>(
&'life0 self,
message: ConsensusMessage<TYPES, LEAF, PROPOSAL>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Send a broadcast message to the entire network.

fn send_event<'life0, 'async_trait>(
&'life0 self,
event: Event<TYPES, LEAF>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Notify the system of an event within hotshot-consensus.

fn public_key(&self) -> &TYPES::SignatureKey

Get a reference to the public key.

fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey

Get a reference to the private key.

fn send_da_broadcast<'life0, 'async_trait, DAPROPOSAL>(
&'life0 self,
message: ConsensusMessage<TYPES, LEAF, DAPROPOSAL>
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
DAPROPOSAL: 'async_trait + ProposalType<NodeType = TYPES>,
Self: 'async_trait,
'life0: 'async_trait,

Send a broadcast to the DA comitee, stub for now

fn is_valid_dac(
&self,
dac: &LEAF::DACertificate,
block_commitment: Commitment<TYPES::BlockType>
) -> bool

Validate a DAC.

fn is_valid_qc(&self, qc: &LEAF::QuorumCertificate) -> bool

Validate a QC.

fn is_valid_vote(
&self,
encoded_key: &EncodedPublicKey,
encoded_signature: &EncodedSignature,
data: VoteData<TYPES, LEAF>,
view_number: TYPES::Time,
vote_token: Checked<TYPES::VoteTokenType>
) -> bool

Validate a vote.

fn accumulate_qc_vote(
&self,
encoded_key: &EncodedPublicKey,
encoded_signature: &EncodedSignature,
leaf_commitment: Commitment<LEAF>,
vote_token: TYPES::VoteTokenType,
view_number: TYPES::Time,
accumlator: CertificateAccumulator<TYPES::VoteTokenType, LEAF>
) -> Either<CertificateAccumulator<TYPES::VoteTokenType, LEAF>, QuorumCertificate<TYPES, LEAF>>

Add a vote for a QC, if the threshold is reached return the QC if not return the accumulator

fn accumulate_da_vote(
&self,
encoded_key: &EncodedPublicKey,
encoded_signature: &EncodedSignature,
block_commitment: Commitment<TYPES::BlockType>,
vote_token: TYPES::VoteTokenType,
view_number: TYPES::Time,
accumlator: CertificateAccumulator<TYPES::VoteTokenType, TYPES::BlockType>
) -> Either<CertificateAccumulator<TYPES::VoteTokenType, TYPES::BlockType>, DACertificate<TYPES>>

Add a vote for a DA QC, if the threshold is reached return the Certificate if not return the accumulator

Provided Methods

fn is_leader<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

returns true if the current node is a leader for the given view_number

fn send_view_error<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time,
error: Arc<HotShotError<TYPES>>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of an error

fn send_replica_timeout<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of a replica timeout

fn send_next_leader_timeout<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

notifies client of a next leader timeout

fn send_decide<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time,
leaf_views: Vec<LEAF>,
decide_qc: LEAF::QuorumCertificate
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

sends a decide event down the channel

fn send_view_finished<'life0, 'async_trait>(
&'life0 self,
view_number: TYPES::Time
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Sends a ViewFinished event

fn sign_da_proposal(
&self,
block_commitment: &Commitment<TYPES::BlockType>
) -> EncodedSignature

Sign a DA proposal.

fn sign_validating_or_commitment_proposal(
&self,
leaf_commitment: &Commitment<LEAF>
) -> EncodedSignature

Sign a validating or commitment proposal.

fn sign_da_vote(
&self,
block_commitment: Commitment<TYPES::BlockType>
) -> (EncodedPublicKey, EncodedSignature)

Sign a vote on DA proposal.

The block commitment and the type of the vote (DA) are signed, which is the minimum amount of information necessary for checking that this node voted on that block.

fn sign_yes_vote(
&self,
leaf_commitment: Commitment<LEAF>
) -> (EncodedPublicKey, EncodedSignature)

Sign a positive vote on validating or commitment proposal.

The leaf commitment and the type of the vote (yes) are signed, which is the minimum amount of information necessary for any user of the subsequently constructed QC to check that this node voted Yes on that leaf. The leaf is expected to be reconstructed based on other information in the yes vote.

fn sign_no_vote(
&self,
leaf_commitment: Commitment<LEAF>
) -> (EncodedPublicKey, EncodedSignature)

Sign a neagtive vote on validating or commitment proposal.

The leaf commitment and the type of the vote (no) are signed, which is the minimum amount of information necessary for any user of the subsequently constructed QC to check that this node voted No on that leaf.

fn sign_timeout_vote(
&self,
view_number: TYPES::Time
) -> (EncodedPublicKey, EncodedSignature)

Sign a timeout vote.

We only sign the view number, which is the minimum amount of information necessary for checking that this node timed out on that view.

This also allows for the high QC included with the vote to be spoofed in a MITM scenario, but it is outside our threat model.

fn create_da_message(
&self,
justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
block_commitment: Commitment<TYPES::BlockType>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, LEAF, PROPOSAL>

Create a message with a vote on DA proposal.

fn create_yes_message(
&self,
justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
leaf_commitment: Commitment<LEAF>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, LEAF, PROPOSAL>

Create a message with a positive vote on validating or commitment proposal.

fn create_no_message(
&self,
justify_qc_commitment: Commitment<LEAF::QuorumCertificate>,
leaf_commitment: Commitment<LEAF>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, LEAF, PROPOSAL>

Create a message with a negative vote on validating or commitment proposal.

fn create_timeout_message(
&self,
justify_qc: LEAF::QuorumCertificate,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, LEAF, PROPOSAL>

Create a message with a timeout votes.